+++ src/filter/LowShelfFilter.cpp
@@ -26,9 +26,9 @@ dibiff::filter::LowShelfFilter::LowShelfFilter(float gain, float cutoff, float s
 dibiff::filter::Coefficients dibiff::filter::LowShelfFilter::calculateCoefficients(float gain, float cutoff, float sampleRate, float qFactor) {
     float A = std::pow(10.0f, gain / 40.0f);
     float w0 = 2.0f * M_PI * cutoff / sampleRate;
-    float alpha = std::sinf(w0) / (2.0f * qFactor);
-    float cosw0 = std::cosf(w0);
-    float sqrtA = std::sqrtf(A);
+    float alpha = std::sin(w0) / (2.0f * qFactor);
+    float cosw0 = std::cos(w0);
+    float sqrtA = std::sqrt(A);
     float b0 = A * ((A + 1.0f) - (A - 1.0f) * cosw0 + 2.0f * sqrtA * alpha);
     float b1 = 2.0f * A * ((A - 1.0f) - (A + 1.0f) * cosw0);
     float b2 = A * ((A + 1.0f) - (A - 1.0f) * cosw0 - 2.0f * sqrtA * alpha);
